requirements_back.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.13.0
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
packaging==21.3
priority==2.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2021.3
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.11
uvicorn==0.18.2
websockets==10.3
wrapt==1.14.1
wsproto==1.0.0
zope.interface==5.4.0
black==22.6.0
flake8==4.0.1
isort==5.10.1
mypy==0.961
mypy-extensions==0.4.3
djangorestframework-stubs==1.7.0
django-stubs==1.12.0
django-stubs-ext==0.5.0